API parity: normalise Mongo updateChild/removeChild return type to int
Closes the last remaining int-vs-long drift from the API parity pass.
Both single-item PATCH primitives now return int on both backends,
matching the JDBC-native shape.

Rationale (user's call): updateFirst modifies at most one document
by definition, and removeChild's (parentId, itemId) query matches
at most one wrapper doc — the returned counts are 0 or 1 in practice.
Casting the driver's long to int at this API boundary is safe.

Diff:
- Tmf630MongoSplitWriteExecutor.updateChild: long → int
  (adds explicit (int) cast on getModifiedCount() return)
- Tmf630MongoSplitWriteExecutor.removeChild: long → int
  (adds explicit (int) cast on getDeletedCount() return)
- Tmf630MongoSplitCollectionIT: local vars long → int, 1L → 1
  in the four assertions on those returns.

Post-pass, all public methods on Tmf630JsonbWriteExecutor and
Tmf630MongoSplitWriteExecutor now have identical signatures.

Reactor build: mvn verify green across all 8 modules.

+++ tmf630-toolkit-mongo-split-collection/src/main/java/org/opentmf/query/tmf630/mongo/split/Tmf630MongoSplitWriteExecutor.java
@@ -228,7 +228,7 @@ public class Tmf630MongoSplitWriteExecutor {
    * the (parent, item) tuple does not exist.
    */
   @Transactional
-  public long updateChild(
+  public int updateChild(
       Class<?> parentType, String parentId, String itemId, Object updatedChild) {
     if (parentId == null) throw new IllegalArgumentException("parentId must not be null");
     if (itemId == null) throw new IllegalArgumentException("itemId must not be null");
@@ -243,7 +243,8 @@ public class Tmf630MongoSplitWriteExecutor {
                 .and(split.itemIdField())
                 .is(itemId));
     Update update = new Update().set(split.payloadField(), payload);
-    return mongoOperations.updateFirst(q, update, split.childCollection()).getModifiedCount();
+    // updateFirst modifies at most one document, so the long return safely narrows.
+    return (int) mongoOperations.updateFirst(q, update, split.childCollection()).getModifiedCount();
   }
 
   /**
@@ -256,7 +257,7 @@ public class Tmf630MongoSplitWriteExecutor {
    * <p>Returns {@code 1} on success, {@code 0} if the child does not exist.
    */
   @Transactional
-  public long removeChild(
+  public int removeChild(
       Class<?> parentType, String parentId, String itemId, Class<?> childType) {
     if (parentId == null) throw new IllegalArgumentException("parentId must not be null");
     if (itemId == null) throw new IllegalArgumentException("itemId must not be null");
@@ -267,7 +268,9 @@ public class Tmf630MongoSplitWriteExecutor {
                 .is(parentId)
                 .and(split.itemIdField())
                 .is(itemId));
-    return mongoOperations.remove(q, split.childCollection()).getDeletedCount();
+    // The (parentId, itemId) predicate matches at most one wrapper, so the long
+    // return safely narrows to int at this API boundary.
+    return (int) mongoOperations.remove(q, split.childCollection()).getDeletedCount();
   }
 
   /**
